requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
apparmor==2.13.4
appdirs==1.4.4
application-utility==1.3.2
asn1crypto==1.3.0
astroid==2.4.1
asttokens==2.0.4
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.2
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
black==19.10b0
boto==2.49.0
boto3==1.14.21
botocore==1.17.21
Brlapi==0.7.0
btrfsutil==1.2.0
CacheControl==0.12.6
cachetools==4.1.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configobj==5.0.6
contextlib2==0.6.0.post1
cryptography==2.9.2
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
distro==1.5.0
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
error-explainer==0.14
fastapi==0.54.2
fasteners==0.15
flake8==3.8.3
future==0.18.2
gast==0.3.3
gensim==3.8.3
Glances==3.1.4.1
google-api-core==1.16.0
google-api-python-client==1.8.3
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
gps==3.20
grpcio==1.30.0
grub2-theme-preview==1.3.2
h11==0.9.0
h5py==2.10.0
html5lib==1.0.1
httplib2==0.18.1
httptools==0.1.1
idna==2.9
ipdb @ file:///home/jacklanda/%E6%A1%8C%E9%9D%A2/ipdb-0.13.2-py3-none-any.whl
ipython @ file:///home/jacklanda/%E6%A1%8C%E9%9D%A2/ipython-7.15.0-py3-none-any.whl
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras-Preprocessing==1.1.2
keyutils==0.6
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lensfun==0.3.95
LibAppArmor==2.13.4
louis==3.14.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
monotonic==1.5
msgpack==1.0.0
mypy==0.770
mypy-extensions==0.4.3
npyscreen==4.10.5
numpy==1.16.1
oauth2client==4.1.3
oauthlib==3.1.0
OpenCC==1.1.1
opencv-python==4.3.0.36
opt-einsum==3.2.1
ordered-set==3.1.1
packaging==20.4
pacman-mirrors==4.16.4
pandas==1.0.5
paramiko==2.7.1
parso==0.7.0
pathspec==0.8.0
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
ply==3.11
powerline-status==2.8.dev9999
prettytable==0.7.2
progress==1.5
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.0
ptyprocess==0.6.0
pwquality==1.4.2
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.7
pycups==2.0.1
pycurl==7.43.0.5
pydantic==1.5.1
PyDrive==1.3.1
pyflakes==2.2.0
Pygments==2.6.1
PyGObject==3.36.1
pylint==2.5.2
pymongo==3.10.1
pymusic-dl==3.0.1
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyserial==3.4
pysmbc==1.0.20
pyspark==3.0.0
pyte==0.8.0
python-dateutil==2.8.1
python-multipart==0.0.5
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
redis==3.5.2
regex==2020.5.14
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rgw==2.0.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.19.2
scipy==1.4.1
Send2Trash==1.5.0
sip==4.19.22
six==1.15.0
sklearn==0.0
smart-open==2.1.0
soupsieve==1.9.5
starlette==0.13.2
team==1.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
thefuck==3.30
thonny==3.2.7
thonny-black-format==0.0.3
thonny-block-highlight==0.3
thonny-error-explainer==0.9
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
udiskie==2.2.0
uritemplate==3.0.1
urllib3==1.25.9
utils==1.0.1
uvicorn==0.11.5
uvloop==0.14.0
variety==0.8.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wordcloud==1.7.0
wrapt==1.12.1
